requirements.txt
Django==1.10.1
django-taggit==0.21.2
django-uuslug==1.1.8
Pillow==3.3.1
psycopg2==2.6.2
python-slugify==1.2.1
Unidecode==0.4.19
wheel==0.24.0
